refactor: moved wallet-precalculation.helper.js to /scripts/helpers
