+++ scripts/helpers/wallet-precalculation.helper.js
